// server-one/src/config/common/enums.ts
export enum NUMBERS {
  ZERO = 0,
  ONE = 1,
  TWO = 2,
  THREE = 3,
  FOUR = 4,
  FIVE = 5,
  SIX = 6,
  SEVEN = 7,
  EIGHT = 8,
  NINE = 9,
  TEN = 10,
}

export enum CONTROLLERS {
  MATRIX = 'MatrixController |',
  AUTH = 'AuthController |',
}

export enum SERVICES {
  MATRIX = 'MatrixService |',
  AUTH = 'AuthService |',
}

export enum REPOSITORIES {
  MATRIX = 'MatrixRepository |',
  AUTH = 'AuthRepository |',
}

export enum METHOD {
  GET = 'GET',
  POST = 'POST',
  PUT = 'PUT',
  DELETE = 'DELETE',
  PATCH = 'PATCH',
}

export enum HTTP_STATUS {
  STATUS_200 = 200,
  STATUS_201 = 201,
  STATUS_204 = 204,
  STATUS_400 = 400,
  STATUS_401 = 401,
  STATUS_403 = 403,
  STATUS_404 = 404,
  STATUS_409 = 409,
  STATUS_415 = 415,
  STATUS_422 = 422,
  STATUS_424 = 424,
  STATUS_429 = 429,
  STATUS_500 = 500,
  STATUS_501 = 501,
  STATUS_503 = 503,
  STATUS_504 = 504,
}

export enum NUMBER {
  ZERO = 0,
  ONE = 1,
  TWO = 2,
  THREE = 3,
  FOUR = 4,
  FIVE = 5,
  SIX = 6,
  SEVEN = 7,
  EIGHT = 8,
  NINE = 9,
  TEN = 10,
}

export enum STATUS_CODE {
  OK = 200,
  CREATED = 201,
  NO_CONTENT = 204,
  BAD_REQUEST = 400,
  UNAUTHORIZED = 401,
  FORBIDDEN = 403,
  NOT_FOUND = 404,
  CONFLICT = 409,
  INTERNAL_SERVER_ERROR = 500,
}